+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to `jupyter-dash` will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.3.0 - ???
+### Added
+ - Added suport for using JupyterDash in Google Colab ([#27](https://github.com/plotly/jupyter-dash/pull/27))
+ - Added support for installing JupyterDash from git using pip: (e.g. `pip install git+https://github.com/plotly/jupyter-dash.git@master`)
+
+### Changed
+ - The default display width in `mode='inline'` is now `100%` to fill the screen width. 
+
 ## 0.2.1 - 2020-05-19
 ### Added
  - Remove f-strings to support Python 3.5